Numerous bug fixes

+++ src/main/kotlin/bot/boobbot/models/SlideShowCommand.kt
@@ -10,6 +10,7 @@ import bot.boobbot.misc.json
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.future.await
 import net.dv8tion.jda.core.EmbedBuilder
+import net.dv8tion.jda.core.Permission
 import net.dv8tion.jda.core.entities.Message
 import java.awt.Color
 
@@ -18,48 +19,23 @@ abstract class SlideShowCommand : AsyncCommand {
 
     private val headers = createHeaders(Pair("Key", BoobBot.config.bbApiKey))
 
+    private val aliases = mapOf(
+        "dick" to "penis",
+        "gif" to "Gifs"
+    )
+    private val allowedEndpoints = arrayOf("boobs", "ass", "dick", "gif", "gay", "tiny", "cumsluts", "collared")
+    private val endpointStr = allowedEndpoints.joinToString(", ")
+
     override suspend fun executeAsync(ctx: Context) {
-        if (ctx.args.isEmpty()) {
+        if (ctx.args.isEmpty() || !allowedEndpoints.contains(ctx.args[0].toLowerCase())) {
             return ctx.embed {
                 setColor(Color.red)
                 setDescription(Formats.error("Missing Args\nbbslideshow <type>\nTypes: boobs, ass, dick, gif, gay, tiny, cumsluts, collared"))
             }
         }
 
-        val endpoint: String
-
-        when (ctx.args[0].toLowerCase()) {
-            "boobs" -> {
-                endpoint = ctx.args[0]
-            }
-            "gay" -> {
-                endpoint = ctx.args[0]
-            }
-            "tiny" -> {
-                endpoint = ctx.args[0]
-            }
-            "cumsluts" -> {
-                endpoint = ctx.args[0]
-            }
-            "collared" -> {
-                endpoint = ctx.args[0]
-            }
-            "ass" -> {
-                endpoint = ctx.args[0]
-            }
-            "dick" -> {
-                endpoint = "penis"
-            }
-            "gif" -> {
-                endpoint = "Gifs"
-            }
-            else -> {
-                return ctx.embed {
-                    setColor(Color.red)
-                    setDescription(Formats.error("What?\nTypes: boobs, ass, dick, gif, gay, tiny, cumsluts, collared"))
-                }
-            }
-        }
+        val query = ctx.args[0].toLowerCase()
+        val endpoint = aliases[query] ?: query
 
         val color = Colors.getEffectiveColor(ctx.message)
         val msg = ctx.sendAsync("\u200B")
@@ -72,7 +48,10 @@ abstract class SlideShowCommand : AsyncCommand {
             delay(5000)
         }
 
-        ctx.message.delete().queue()
+        if (ctx.botCan(Permission.MESSAGE_MANAGE)) {
+            ctx.message.delete().queue()
+        }
+
         msg.delete().queue()
     }
